Rename esbuild-options to esbuild, export parseArgs and touchManifest function
